Move Parser::Create* methods to AstFactory.cpp
Move all the ParseNode creation methods from Parse.cpp to a new file
named AstFactory.cpp.  The methods are unchanged.  This is in
preparation for moving these methods out of the Parser class to
declutter the class and make it easier to read and digest both the
Parser and the ParseNode creation logic at a glance.