Eliminate IdentToken and buildAST bifurcated logic
Deferred parse and non-deferred parse should behave the same as much as
possible for correct parity between the two modes.  This is best
accomplished by using the same mechanisms in both for verifying the
grammar and relying as little as possible on the AST nodes to do so
(since deferred parse does not have (all of) them).

For various parse time checks and tracking logic, i.e. MarkEscapingRef,
CheckStrictModeEvalArgumentsUsage, and TrackAssignment, we can always
use the IdentToken term passed to ParseExpr and ParseTerm for both parse
modes instead of using a ParseNode for buildAST and the IdentToken for
!buildAST.

To make this work ParseTerm, ParseExpr, and others are all updated to
always populate the IdentToken out parameter regardless of the buildAST
template parameter.

Further, we can use IdentPtr instead of IdentToken since IdentToken is
only used for tracking knopName/tkID pids, and its min/lim fields are no
longer used.

Parser::NodeIsEvalName() is no longer used after these changes and has
also been removed.