Simplify spread parsing
Treat `...` as punctuation and not an operator.  `...` is not an
operator, but was treated as such because it was one way to deal with
arrow function parsing, where arrow functions allow `...` but
parenthetical expressions don't, but we don't know we have an arrow
function until we see the `=>`, so we have to defer erroring on `...`
until we know we do not have an arrow function.

This solution incurred a bit of state and bookkeeping on the parser and
is awkward from a grammar point of view since `...` is not an operator.

The bookkeeping can be reduced to on stack context passed between
ParseTerm and ParseExpr only by treating `...` as punctuation only in
the grammar productions where it appears. Tracking parentheses count is
not needed.  All that is needed is whether an ellipsis was seen, and its
location for error reporting.